solution #7

+++ app/CarWashStation.py
@@ -5,7 +5,7 @@ from typing import List
 class CarWashStation:
     def __init__(
         self,
-        distance_from_city_center: float,
+        distance_from_city_center: int,
         clean_power: int,
         average_rating: float,
         count_of_ratings: int,
